+++ core/src/dird/authenticate_console.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2001-2008 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2019 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -216,8 +216,6 @@ bool ConsoleAuthenticatorFrom_18_2::SendResponseMessage(uint32_t response_id,
 bool ConsoleAuthenticatorFrom_18_2::SendInfoMessage()
 {
   std::string message;
-  message += kBareosVersionStrings.BinaryInfo;
-  message += " binary\n";
   message += kBareosVersionStrings.ServicesMessage;
   message += "\n";
   message += "You are ";